updated with more details and also added examples for some core java api

+++ src/casting/ExplicitCast.java
@@ -6,8 +6,9 @@ public class ExplicitCast {
 
         //Without explicit cast compiler suspects overflow hence compile error.
         //Overflow situation because literal, 1.0, value is treated as double data type.
-        //Fix: float size = (float) 1.0;
-        float size = 1.0;
+        //Compile Error:
+        // float size = 1.0;
+        float size = (float) 1.0;
         System.out.println("Explicit cast size value : " + size);
 
     }
